+++ test/src/test/java/com/lkn/game/B.java
@@ -62,7 +62,7 @@ public class B extends Shape {
     private final byte[][] shape8 = genericArr(origin8);
     private final int[] posArr8 = genericPosArr(origin8);
 
-    public B(byte[][] board) {
+    public B(int[][] board) {
         super(board);
     }
