+++ test/src/test/java/com/lkn/game/L.java
@@ -34,7 +34,7 @@ public class L extends Shape {
     private final byte[][] shape4 = genericArr(origin4);
     private final int[] posArr4 = genericPosArr(origin4);
 
-    public L(byte[][] board) {
+    public L(int[][] board) {
         super(board);
     }
